+++ src/scipp/core/cumulative.py
@@ -5,19 +5,21 @@
 from typing import Literal, Optional
 
 from .._scipp import core as _cpp
+from ..typing import VariableLikeType
 from ._cpp_wrapper_util import call_func as _call_cpp_func
+from .concepts import transform_data
 
 
-def cumsum(a: _cpp.Variable,
+def cumsum(a: VariableLikeType,
            dim: Optional[str] = None,
-           mode: Literal['exclusive', 'inclusive'] = 'inclusive') -> _cpp.Variable:
+           mode: Literal['exclusive', 'inclusive'] = 'inclusive') -> VariableLikeType:
     """Return the cumulative sum along the specified dimension.
 
     See :py:func:`scipp.sum` on how rounding errors for float32 inputs are handled.
 
     Parameters
     ----------
-    a:
+    a: scipp.typing.VariableLike
         Input data.
     dim:
         Optional dimension along which to calculate the sum. If not
@@ -28,10 +30,14 @@ def cumsum(a: _cpp.Variable,
 
     Returns
     -------
-    :
+    : Same type as input
         The cumulative sum of the input values.
     """
-    if dim is None:
-        return _call_cpp_func(_cpp.cumsum, a, mode=mode)
-    else:
-        return _call_cpp_func(_cpp.cumsum, a, dim, mode=mode)
+
+    def _cumsum(data):
+        if dim is None:
+            return _call_cpp_func(_cpp.cumsum, data, mode=mode)
+        else:
+            return _call_cpp_func(_cpp.cumsum, data, dim, mode=mode)
+
+    return transform_data(a, _cumsum)