+++ tests/hypothesis/bin_test.py
@@ -1,8 +1,10 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
 # @author Simon Heybrock
+import itertools
 from hypothesis import given, settings
 from hypothesis import strategies as st
+from scipp.testing import strategies as scst
 
 import scipp as sc
 import numpy as np
@@ -41,3 +43,29 @@ def test_automatic_grouping_optimization(nrow, i, j):
     assert sc.identical(
         table.group('label').coords['label'],
         sc.array(dims=['label'], values=np.unique(table.coords['label'].values)))
+
+
+@settings(max_examples=100, deadline=1000)
+@given(st.integers(min_value=1, max_value=12345),
+       scst.variables(dtype=bool, ndim=st.integers(min_value=1, max_value=3)))
+def test_bins_concat_masking(nevent, mask):
+    from numpy.random import default_rng
+    rng = default_rng(seed=1234)
+    mask.unit = None
+    table = sc.DataArray(sc.arange('row', nevent))
+    for dim in mask.dims:
+        table.coords[dim] = sc.array(dims=['row'], values=rng.random((nevent, )))
+    da = table.bin(mask.sizes)
+    for n in range(1, mask.ndim + 1):
+        for dims in itertools.permutations(mask.dims, n):
+            m = mask
+            for dim in mask.dims:
+                if dim not in dims:
+                    m = m[dim, 0]
+            da.masks["".join(dims)] = m
+    for n in range(1, da.ndim + 1):
+        for dims in itertools.permutations(da.dims, n):
+            assert sc.identical(da.bins.concat(dims).hist(), da.hist().sum(dims))
+            assert sc.identical(da.transpose().bins.concat(dims).hist(),
+                                da.transpose().hist().sum(dims))
+    assert sc.identical(da.bins.concat(None).hist(), da.hist().sum(None))